% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_create}
\alias{box_create}
\title{Create a new box}
\usage{
box_create(name, activate = TRUE)
}
\arguments{
\item{name}{box name.}

\item{activate}{Activate box after creation? Default is \code{TRUE}.}
}
\description{
Create a new box
}
