% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_delete}
\alias{box_delete}
\title{Delete box}
\usage{
box_delete(name, ..., .skip = TRUE)
}
\arguments{
\item{name}{box name(s) to delete.}

\item{...}{More comma separated box names (optional).}

\item{.skip}{Skip box if it does not exist (default is \code{TRUE}).}
}
\description{
Delete one or more boxes. If a box does not exist it is skipped.
}
\examples{
\dontrun{
box_delete("a")
box_delete(c("a", "b"))
box_delete("a", "b") # same as above
}
}
