% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_export}
\alias{box_export}
\title{Export box to a file}
\usage{
box_export(name, file = NULL)
}
\arguments{
\item{name}{box name.}

\item{file}{Where to store the box. We suggest the file ending (`.box`).
If `NULL`, the box is exported to the current working dir.}
}
\description{
The exported box file contains the box name, the original box path and its name and the databse itself.
}
