% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_import}
\alias{box_import}
\title{Import box from file}
\usage{
box_import(file, name = NULL, overwrite = FALSE, activate = FALSE)
}
\arguments{
\item{file}{Path to box file.}

\item{name}{box name to use for imported data. If `NULL`, the original name (before export) will be used.}

\item{overwrite}{Overwrite if box with name already exists?}

\item{activate}{Activate box after import?}
}
\description{
Import box from file
}
