% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_purge}
\alias{box_purge}
\title{Purge (expired) box items}
\usage{
box_purge(name = NULL, cutoff_date = Sys.time(), purge_all = FALSE)
}
\arguments{
\item{name}{Name of box.}

\item{cutoff_date}{Threshold (date / datetime) for expiration. If \code{NULL} (default) the current time is used.}

\item{purge_all}{Remove all items? If \code{FALSE} (default), only expired items (before \code{cutoff_date}) are removed.}
}
\description{
By default, removes items that have expired from active box.
}
