% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxed-package.R
\docType{package}
\name{boxed-package}
\alias{boxed}
\alias{boxed-package}
\title{boxed: Easily store and retrieve R objects}
\description{
Useful tools to easily store any R objects and retrieve them again later.
}
\keyword{internal}
