% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{depot}
\alias{depot}
\title{Show depot storage content}
\usage{
depot(name = NULL)
}
\arguments{
\item{name}{Depot names. If `NULL`, the active depot is used (`depot_active()`).}
}
\value{
Tibble with depot content.
}
\description{
Show depot storage content
}
