% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{depot_create}
\alias{depot_create}
\title{Create a new depot}
\usage{
depot_create(name, activate = TRUE)
}
\arguments{
\item{name}{Depot name.}

\item{activate}{Activate depot after creation? Default is `TRUE`.}
}
\description{
Create a new depot
}
