% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{depot_delete}
\alias{depot_delete}
\title{Delete depot}
\usage{
depot_delete(name, ..., .skip = TRUE)
}
\arguments{
\item{name}{Depot name(s) to delete.}

\item{...}{More comma separated depot names (optional).}

\item{.skip}{Skip depot if it does not exist (default is `TRUE`).}
}
\description{
Delete one or more depots. If a depot does not exist it is skipped.
}
\examples{
\dontrun{
  depot_delete("a")
  depot_delete(c("a", "b"))
  depot_delete("a", "b")  # same as above
}
}
