% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{depot_export}
\alias{depot_export}
\title{Export depot to a file}
\usage{
depot_export(name, file = NULL)
}
\arguments{
\item{name}{Depot name.}

\item{file}{Where to store the depot. We suggest the file ending (`.depot`).
If `NULL`, the depot is exported to the current working dir.}
}
\description{
The exported depot file contains the depot name, the original depot path and its name and the databse itself.
}
