% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{depot_import}
\alias{depot_import}
\title{Import depot from file}
\usage{
depot_import(file, name = NULL, overwrite = FALSE, activate = FALSE)
}
\arguments{
\item{file}{Path to depot file.}

\item{name}{Depot name to use for imported data. If `NULL`, the original name (before export) will be used.}

\item{overwrite}{Overwrite if depot with name already exists?}

\item{activate}{Activate depot after import?}
}
\description{
Import depot from file
}
