% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{depot_path}
\alias{depot_path}
\title{Path to depot database file on disk}
\usage{
depot_path(name = NULL)
}
\arguments{
\item{name}{Depot names. If `NULL`, the active depot is used (`depot_active()`).}
}
\description{
Path to depot database file on disk
}
