% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{depot_size}
\alias{depot_size}
\title{Number of objects in depot}
\usage{
depot_size(name = NULL)
}
\arguments{
\item{name}{Depot names. If `NULL`, the active depot is used (`depot_active()`).}
}
\value{
Named vector with number of objects per depot.
}
\description{
Number of objects in depot
}
