% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{.depot_connection}
\alias{.depot_connection}
\title{Create depot database connection}
\usage{
.depot_connection(name, must_exist = FALSE)
}
\arguments{
\item{name}{Depot name.}

\item{must_exist}{Security mechanism. Will prevent creation of databse if it does not already exists.}
}
\value{
Connection object.
}
\description{
Create depot database connection
}
\keyword{internal}
