% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{item_pick}
\alias{item_pick}
\alias{pick}
\title{Get object from box}
\usage{
item_pick(id, box = NULL)

pick(id, box = NULL)
}
\arguments{
\item{id}{Object id}

\item{box}{box name. If `NULL`, the active box is used (`box_active()`).}
}
\description{
Get object from box
}
