% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{item_remove}
\alias{item_remove}
\alias{remove}
\title{Remove an item from a box}
\usage{
item_remove(id, box = NULL)

remove(id, box = NULL)
}
\arguments{
\item{id}{Item id.}

\item{box}{Box name. If \code{NULL}, the active box is used (\code{box_active()}).}
}
\value{
\code{TRUE} if deletion succeeded, else \code{FALSE}.
}
\description{
Remove an item from a box
}
