% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\title{Keep object for later use.}
\usage{
keep(id, obj, info = NULL, tags = NULL)
}
\arguments{
\item{id}{Unique ID of storage slot.}

\item{obj}{Object to store.}

\item{info}{Some information about the object.}

\item{tags}{One or more tags (character vector or comma separated string)}
}
\description{
Keep object for later use.
}
