% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep_file}
\alias{keep_file}
\title{Keep object for later use.}
\usage{
keep_file(id, path, info = NULL, tags = NULL)
}
\arguments{
\item{id}{Unique ID of storage slot.}

\item{path}{File path.}

\item{info}{Some information about the object.}

\item{tags}{One or more tags (character vector or comma separated string)}
}
\description{
Keep object for later use.
}
\examples{
file <- system.file("ext/depot_init.sql", package = "keeper")
keep_file("code", file, "some SQL code")
}
