% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keeper-package.R
\docType{package}
\name{keeper-package}
\alias{keeper}
\alias{keeper-package}
\title{keeper: Easily keep and retrieve R objects}
\description{
Useful tools to easily store any R objects and retrieve them again later.
}
\keyword{internal}
