% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{pack_file}
\alias{pack_file}
\title{Pack file into a box.}
\usage{
pack_file(id, path, info = NULL, tags = NULL)
}
\arguments{
\item{id}{Unique ID of storage slot.}

\item{path}{File path.}

\item{info}{Some information about the object.}

\item{tags}{One or more tags (character vector or comma separated string)}
}
\description{
Pack file into a box.
}
\examples{
file <- system.file("ext/box_init.sql", package = "boxes")
pack_file("code", file, "some SQL code")
}
