% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{print.box_df}
\alias{print.box_df}
\title{Print a box dataframe}
\usage{
\method{print}{box_df}(x, ...)
}
\arguments{
\item{x}{A \code{box_df} object.}

\item{...}{Passed on.}
}
\description{
Adds the active box at top before printing dataframe.
}
\keyword{internal}
