% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box}
\alias{box}
\title{Show box content}
\usage{
box(name = NULL)
}
\arguments{
\item{name}{box names. If `NULL`, the active box is used (`box_active()`).}
}
\value{
Tibble with box content.
}
\description{
Show box content
}
