% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_path}
\alias{box_path}
\title{Path to box database file on disk}
\usage{
box_path(name = NULL)
}
\arguments{
\item{name}{box names. If \code{NULL}, the active box is used (\code{box_active()}).}
}
\description{
Path to box database file on disk
}
