% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_size}
\alias{box_size}
\title{Number of objects in box}
\usage{
box_size(name = NULL)
}
\arguments{
\item{name}{box names. If `NULL`, the active box is used (`box_active()`).}
}
\value{
Named vector with number of objects per box.
}
\description{
Number of objects in box
}
