% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxed-package.R
\docType{package}
\name{boxes-package}
\alias{boxes-package}
\title{boxes: Easily store and retrieve R objects}
\description{
Useful tools to easily store any R objects and retrieve them again later.
}
\keyword{internal}
