% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{.box_connection}
\alias{.box_connection}
\title{Create box database connection}
\usage{
.box_connection(name = NULL, must_exist = TRUE)
}
\arguments{
\item{name}{box names. If \code{NULL}, the active box is used (\code{box_active()}).}

\item{must_exist}{Security mechanism. If \code{TRUE} (default), prevents creating a database with the given name, if it
does not yet exist.}
}
\value{
Connection object.
}
\description{
Create box database connection
}
\keyword{internal}
