% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{.depot_connection}
\alias{.depot_connection}
\title{Create depot database connection}
\usage{
.depot_connection(name = NULL, must_exist = TRUE)
}
\arguments{
\item{name}{Depot names. If `NULL`, the active depot is used (`depot_active()`).}

\item{must_exist}{Security mechanism. If `TRUE` (default), prevents creating a database with the given name, if it
does not yet exist.}
}
\value{
Connection object.
}
\description{
Create depot database connection
}
\keyword{internal}
