% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{item_pack}
\alias{item_pack}
\alias{pack}
\title{Pack item into a box.}
\usage{
item_pack(
  obj,
  id = NULL,
  info = NULL,
  tags = NULL,
  box = NULL,
  replace = FALSE
)

pack(obj, id = NULL, info = NULL, tags = NULL, box = NULL, replace = FALSE)
}
\arguments{
\item{obj}{Object to store.}

\item{id}{Unique id as storage name. If \code{NULL}, the objects's name is used.}

\item{info}{Some information about the object.}

\item{tags}{One or more tags (character vector or comma separated string).}

\item{box}{Name of box to use. If \code{NULL}, the active box used (\code{box_active()}).}

\item{replace}{Replace object if \code{id} already exists (default \code{FALSE}).}
}
\value{
Returns \code{id}.
}
\description{
Pack item into a box.
}
