% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{item_pick}
\alias{item_pick}
\alias{pick}
\title{Get object from box}
\usage{
item_pick(id, box = NULL, remove = FALSE)

pick(id, box = NULL, remove = FALSE)
}
\arguments{
\item{id}{Object id}

\item{box}{Box name. If `NULL`, the active box is used (`box_active()`).}

\item{remove}{Remove object from box when retrieving it? (default `FALSE`)}
}
\description{
Get object from box
}
