% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{item_pick}
\alias{item_pick}
\alias{pick}
\title{Get object from box}
\usage{
item_pick(id = NULL, box = NULL, remove = FALSE)

pick(id = NULL, box = NULL, remove = FALSE)
}
\arguments{
\item{id}{Object id. If \code{NULL} (default), the latest item is picked.}

\item{box}{Box name. If \code{NULL}, the active box is used (\code{box_active()}).}

\item{remove}{Remove object from box when retrieving it? (default \code{FALSE})}
}
\description{
Get object from box
}
