% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{item_remove}
\alias{item_remove}
\alias{remove}
\title{Remove an item from a box}
\usage{
item_remove(id, name = NULL)

remove(id, name = NULL)
}
\arguments{
\item{id}{Item id.}

\item{name}{Box name. If `NULL`, the active box is used (`box_active()`).}
}
\value{
`TRUE` if deletion succeeded, else `FALSE`.
}
\description{
Remove an item from a box
}
