% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{pick}
\alias{pick}
\title{Get object from depot}
\usage{
pick(id, depot = NULL)
}
\arguments{
\item{id}{Object id}

\item{depot}{Depot name. If `NULL`, the active depot is used (`depot_active()`).}
}
\description{
Get object from depot
}
